Rename script directory, add documentation and clean up
