+++ scripts/package-versions/index.js
@@ -1,0 +1,193 @@
+/**
+ * @file Increments Cloud Manager package versions.
+ *
+ * Usage:
+ *
+ * package-versions <manager-version> <api-version> <validation-version> [-f | --force]
+ *
+ * Positional Parameters:
+ * - `<manager-version>`    (Optional) Desired Cloud Manager package version.
+ * - `<api-version>`        (Optional) Desired APIv4 package version.
+ * - `<validation-version>` (Optional) Desired Validation package version.
+ *
+ * Optional Flags:
+ * - `-f | --force`         Forces the script to update package versions without
+ *                          prompting the user for confirmation. When this is
+ *                          specified, the script expects versions for all
+ *                          packages to be passed as positional arguments, and
+ *                          will fail otherwise.
+ */
+
+const fs = require('fs');
+const path = require('path');
+const readline = require('readline');
+
+// Command line arguments with executable path and name excluded.
+const args = process.argv.slice(2);
+
+/**
+ * Package versions that were passed to CLI.
+ *
+ * @var {string[]}
+ */
+const desiredVersions = args.filter((arg) => {
+  return !arg.startsWith('-');
+});
+
+/**
+ * CLI flags that modify script behavior.
+ *
+ * @var {string[]}
+ */
+const flags = args.filter((arg) => {
+  return arg.startsWith('-');
+});
+
+/**
+ * Path to repository root.
+ *
+ * @var {string}
+ */
+const root = path.resolve(__dirname, '..', '..');
+
+/**
+ * Gets the path to the package.json file for the package with the given name.
+ *
+ * @param {string} packageName - Name of package for which to retrieve path.
+ *
+ * @returns {string} Package path for `packageName`.
+ */
+const getPackagePath = (packageName) => {
+  return path.join(root, 'packages', packageName, 'package.json');
+};
+
+/**
+ * Reads input from stdin and returns a Promise that resolves to the input value.
+ *
+ * @return {Promise<string>}
+ */
+const readlinePromise = (prompt) => {
+  const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout
+  });
+
+  return new Promise((resolve) => {
+    rl.question(prompt, (result) => {
+      rl.close();
+      resolve(result);
+    });
+  });
+}
+
+/**
+ * If `true`, packages will be updated using desired versions passed via CLI.
+ *
+ * In this case, the user is expected to have passed a version number for each
+ * package and will not be prompted to specify versions interactively. If the
+ * user fails to provide a version for one of the packages, the script exit with
+ * an error.
+ *
+ * @var {boolean}
+ */
+const force = flags.includes('-f') || flags.includes('--force');
+
+/**
+ * Desired package versions from command line arguments.
+ */
+const [
+  desiredManagerVersion,
+  desiredApiVersion,
+  desiredValidationVersion,
+] = desiredVersions;
+
+// Describes packages that should be modified by this script.
+const jobs = [
+  { name: 'manager', path: getPackagePath('manager'), desiredVersion: desiredManagerVersion },
+  { name: 'api-v4', path: getPackagePath('api-v4'), desiredVersion: desiredApiVersion },
+  { name: 'validation', path: getPackagePath('validation'), desiredVersion: desiredValidationVersion },
+];
+
+// Describes changes that were made by this script. Used to summarize changes
+// to the user.
+const summary = [];
+
+/**
+ *
+ */
+const main = async () => {
+  const currentPackageInfo = jobs.map((job) => {
+    return {
+      name: job.name,
+      currentVersion: JSON.parse(fs.readFileSync(job.path, 'utf8')).version || '(No Version)',
+    };
+  });
+
+  console.info('Package info:');
+  console.table(currentPackageInfo);
+
+  for (job of jobs) {
+    const jobName = job.name;
+    const jobPath = job.path;
+    const jobVersion = job.desiredVersion;
+
+    const packageData = JSON.parse(fs.readFileSync(jobPath));
+    const currentVersion = packageData.version;
+
+    let result = undefined;
+    if (!force) {
+      const suggestedVersion = jobVersion ? jobVersion : (() => {
+        // If package doesn't have a version for some reason, suggest `0.1.0`.
+        if (!currentVersion) {
+          return `0.1.0`;
+        }
+        // Assumes `major.minor.patch` format, undefined behavior otherwise.
+        const [major, minor, patch] = currentVersion.split('.');
+        return `${major || 0}.${Number(minor || 0) + 1}.${patch || 0}`;
+      })();
+
+      const prompt = `New version for '${jobName}' package? (${ suggestedVersion })\n`;
+      result = await readlinePromise(prompt);
+      if (!result) {
+        result = suggestedVersion;
+      }
+    }
+    else {
+      if (!jobVersion) {
+        throw new Error(`Unable to increment package '${jobName}' version; -y flag was passed but no version was specified for this package`);
+      }
+      console.info(`Incrementing package '${jobName}' version to '${jobVersion}'`);
+      result = jobVersion;
+    }
+
+    packageData.version = result;
+    const jsonString = `${JSON.stringify(packageData, null, 2)}\n`;
+    fs.writeFileSync(jobPath, jsonString);
+    summary.push({
+      name: jobName,
+      oldVersion: currentVersion,
+      newVersion: result,
+    });
+  };
+};
+
+// Run program, display errors, summary, etc .
+(async () => {
+  let error = false;
+  try {
+    await main();
+    console.info('Successfully incremented package versions. See the summary below:');
+  }
+  catch (e) {
+    console.error(e.message || 'An unknown error has occurred.');
+    console.error('An error has occurred and package versions were not incremented as expected.')
+    console.info('Refer to the summary below to see what changes were made.');
+    error = true;
+  }
+  finally {
+    console.table(summary);
+    if (error) {
+      process.exit(1);
+    }
+  }
+})();
